// src/enums/site-type.enums.ts
export enum SiteTypeRefs {
  SITE_ID = 'Site ID',
  ENGLISH_NAME = 'English name',
  FRENCH_NAME = 'French name',
  KHMER_NAME = 'Khmer name',
  ALTERNATIVE_NAME = 'Alternative name',
  ALTERNATIVE_KHMER_NAME = 'Alternative Khmer name',
  DESCRIPTION = 'Description',
  IK_ID = 'IK ID',
  MH_ID = 'MH ID',
  RESEARCHERS = 'Researchers',
  VERIFIED = 'Verified',
  VERIFICATION_DATE = 'Verification date',
  LOCATED_BY = 'Located by',
  CERAMICS_DETAILS = 'Ceramics details',
  CERAMICS = 'Ceramics',
  BUILD_MATERIALS = 'Build materials',
  BUILD_MATERIAL_COMMENTS = 'Build material comments',
  ARTEFACT_COMMENTS = 'Artefacts comments',
  LOOTED = 'Looted',
  CULTIVATED = 'Cultivated',
  CLEARED = 'Cleared',
  THREATENED = 'Threatened',
  DATABASING_COMMENTS = 'Databasing comments',
  CREATION_DATE = 'Creation date',
  MODIFICATION_DATE = 'Modification date',
  USER_CREATION = 'Created by',
  USER_MODIFICATION = 'Modified by',
  FEATURE_TYPE = 'Feature type',
  STUDY_AREA = 'Study area',
  ARTEFACTS = 'Artefacts',
}

export enum SiteAttributes {
  SITE_ID = 'siteId',
  ENGLISH_NAME = 'englishName',
  FRENCH_NAME = 'frenchName',
  KHMER_NAME = 'khmerName',
  ALTERNATIVE_NAME = 'alternativeName',
  ALTERNATIVE_KHMER_NAME = 'alternativeKhmerName',
  DESCRIPTION = 'description',
  VERIFIED = 'verified',
  VERIFICATION_DATE = 'verificationDate',
  LOCATED_BY = 'locatedBy',
  CERAMICS_DETAILS = 'ceramicsDetails',
  CERAMICS = 'ceramics',
  BUILD_MATERIAL_COMMENTS = 'buildMaterialComments',
  ARTEFACTS_COMMENTS = 'artefactsComments',
  LOOTED = 'looted',
  CULTIVATED = 'cultivated',
  CLEARED = 'cleared',
  THREATENED = 'threatened',
  DATABASING_COMMENTS = 'databasingComments',
  CREATION_DATE = 'creationDate',
  MODIFICATION_DATE = 'modificationDate',
  USER_CREATION = 'userCreation',
  USER_MODIFICATION = 'userModification',
  STUDY_AREA = 'studyArea',
  FEATURE_TYPE = 'featureType',
  DOCUMENTS = 'documents',
  ARTEFACTS = 'artefacts',
  INDIVIDUALS = 'individuals',
  GROUND_VERIFIED_BY = 'groundVerifiedBy',
  MATERIAL_SANDSTONE = 'materialSandstone',
  MATERIAL_PINK_SANDSTONE = 'materialPinkSandstone',
  MATERIAL_LATERITE = 'materialLaterite',
  MATERIAL_BRICK = 'materialBrick',
  MATERIAL_THMAPHNOM = 'materialThmaphnom',
  MATERIAL_OTHER = 'materialOther',
  ARTEFACTS_PEDESTAL = 'artefactsPedestal',
  ARTEFACTS_COLONETTE = 'artefactsColonette',
  ARTEFACTS_DOORFRAME = 'artefactsDoorframe',
  ARTEFACTS_STATUARY = 'artefactsStatuary',
  ARTEFACTS_DECORATED_BLOCKS = 'artefactsDecoratedBlocks',
  ARTEFACTS_OTHER = 'artefactsOther',
  ID_AMBIGUITY = 'idAmbiguity',
  ID_AMBIGUITY_DETAIL = 'idAmbiguityDetail',
  IK_ID = 'ikId',
  MH_ID = 'mhId',
  CK_ID = 'ckId',
  IK_STARRED = 'ikStarred',
  BD_ID = 'bdId',
  GT_ID = 'gtId',
  CP_ID = 'cpId',
  DC_ID = 'dcId',
  GA_ID = 'gaId',
  JB_ID = 'jbId',
  EM_ID = 'emId',
  PK_ID = 'pkId',
  IS_ID = 'isId',
  COORDINATE_SOURCE = 'coordinateSource',
  IN_GAP_STUDY_AREA = 'inGapStudyArea',
  IN_LIDAR_STUDY_AREA = 'inLidarStudyArea',
  TO_BE_INVESTIGATED = 'toBeInvestigated',
  TO_BE_REMAPPED = 'toBeRemapped',
  TO_BE_REMOVED = 'toBeRemoved',
  DB_RESOLVED = 'dbResolved',
}